+++ src/behaviour/Collision.js
@@ -14,10 +14,11 @@ export default class Collision extends Behaviour {
    * @param {function} onCollide - Function to call when particles collide
    * @param {number} life - The life of the particle
    * @param {function} easing - The behaviour's decaying trend
+   * @param {boolean} [isEnabled=true] - Determines if the behaviour will be applied or not
    * @return void
    */
-  constructor(emitter, useMass, onCollide, life, easing) {
-    super(life, easing, type);
+  constructor(emitter, useMass, onCollide, life, easing, isEnabled = true) {
+    super(life, easing, type, isEnabled);
 
     this.reset(emitter, useMass, onCollide);
   }
@@ -51,7 +52,7 @@ export default class Collision extends Behaviour {
    * @param {integer} index - the particle index
    * @return void
    */
-  applyBehaviour(particle, time, index) {
+  mutate(particle, time, index) {
     const particles = this.emitter
       ? this.emitter.particles.slice(index)
       : this.particles.slice(index);