+++ src/emitter/Emitter.js
@@ -2,12 +2,12 @@ import {
   DEFAULT_BIND_EMITTER,
   DEFAULT_BIND_EMITTER_EVENT,
   DEFAULT_DAMPING,
-  DEFAULT_EMITTER_RATE
+  DEFAULT_EMITTER_RATE,
 } from './constants';
 import EventDispatcher, {
   PARTICLE_CREATED,
   PARTICLE_DEAD,
-  PARTICLE_UPDATE
+  PARTICLE_UPDATE,
 } from '../events';
 import { INTEGRATION_TYPE_EULER, integrate } from '../math';
 
@@ -26,7 +26,7 @@ export default class Emitter extends Particle {
   /**
    * Constructs an Emitter instance.
    *
-   * @param {object} properties - The properties to instantiate the particle with
+   * @param {object} properties - The properties to instantiate the emitter with
    * @return void
    */
   constructor(properties) {
@@ -55,6 +55,12 @@ export default class Emitter extends Particle {
      * @type {array}
      */
     this.behaviours = [];
+
+    /**
+     * @desc The behaviours for the emitter.
+     * @type {array}
+     */
+    this.emitterBehaviours = [];
 
     /**
      * @desc The current emit iteration.
@@ -142,6 +148,21 @@ export default class Emitter extends Particle {
     const { x = position.x, y = position.y, z = position.z } = newPosition;
 
     this.position.set(x, y, z);
+
+    return this;
+  }
+
+  /**
+   * Sets the rotation of the emitter.
+   *
+   * @param {object} newRotation - an object the new x, y and z props
+   * @return {Emitter}
+   */
+  setRotation(newRotation = {}) {
+    const { rotation } = this;
+    const { x = rotation.x, y = rotation.y, z = rotation.z } = newRotation;
+
+    this.rotation.set(x, y, z);
 
     return this;
   }
@@ -330,6 +351,81 @@ export default class Emitter extends Particle {
   }
 
   /**
+   * Adds an emitter behaviour to the emitter.
+   *
+   * @param {Behaviour} behaviour - The behaviour to add to the emitter
+   * @return {Emitter}
+   */
+  addEmitterBehaviour(behaviour) {
+    this.emitterBehaviours.push(behaviour);
+
+    behaviour.initialize(this);
+
+    return this;
+  }
+
+  /**
+   * Adds multiple behaviours to the emitter.
+   *
+   * @param {array<Behaviour>} behaviours - an array of emitter behaviours
+   * @return {Emitter}
+   */
+  addEmitterBehaviours(behaviours) {
+    let i = behaviours.length;
+
+    while (i--) {
+      this.addEmitterBehaviour(behaviours[i]);
+    }
+
+    return this;
+  }
+
+  /**
+   * Sets the emitter's behaviours.
+   *
+   * @param {array<Behaviour>} behaviours - an array of emitter behaviours
+   * @return {Emitter}
+   */
+  setEmitterBehaviours(behaviours) {
+    const length = behaviours.length;
+
+    this.emitterBehaviours = behaviours;
+
+    for (let i = 0; i < length; i++) {
+      this.emitterBehaviours[i].initialize(this);
+    }
+
+    return this;
+  }
+
+  /**
+   * Removes the behaviour from the emitter's behaviours array.
+   *
+   * @param {Behaviour} behaviour - The behaviour to remove
+   * @return {Emitter}
+   */
+  removeEmitterBehaviour(behaviour) {
+    const index = this.emitterBehaviours.indexOf(behaviour);
+
+    if (index > -1) {
+      this.emitterBehaviours.splice(index, 1);
+    }
+
+    return this;
+  }
+
+  /**
+   * Removes all behaviours from the emitter.
+   *
+   * @return {Emitter}
+   */
+  removeAllEmitterBehaviours() {
+    Util.destroyArray(this.emitterBehaviours);
+
+    return this;
+  }
+
+  /**
    * Creates a particle by retreiving one from the pool and setting it up with
    * the supplied initializer and behaviour.
    *
@@ -353,8 +449,7 @@ export default class Emitter extends Particle {
    * @return void
    */
   setupParticle(particle) {
-    var initializers = this.initializers;
-    var behaviours = this.behaviours;
+    const { initializers, behaviours } = this;
 
     InitializerUtil.initialize(this, particle, initializers);
 
@@ -395,6 +490,26 @@ export default class Emitter extends Particle {
         this.parent.pool.expire(particle.reset());
         this.particles.splice(i, 1);
       }
+    }
+
+    this.updateEmitterBehaviours(time);
+  }
+
+  /**
+   * Updates the emitter's emitter behaviours.
+   *
+   * @param {number} time - Proton engine time
+   * @return void
+   */
+  updateEmitterBehaviours(time) {
+    if (this.sleep) {
+      return;
+    }
+
+    const length = this.emitterBehaviours.length;
+
+    for (let i = 0; i < length; i++) {
+      this.emitterBehaviours[i].applyBehaviour(this, time, i);
     }
   }