+++ test/core/fromJSON.spec.js
@@ -4,9 +4,11 @@ import * as Proton from '../../src';
 
 import Particles from '../../src/core/Proton';
 import chai from 'chai';
+import containsEmitterBehaviours from './fixtures/json/containsEmitterBehaviours.json';
 import containsInvalidBehaviour from './fixtures/json/containsInvalidBehaviour.json';
 import containsInvalidInitializer from './fixtures/json/containsInvalidInitializer.json';
 import containsInvalidZoneType from './fixtures/json/containsInvalidZoneType.json';
+import containsTotalEmitTimes from './fixtures/json/containsTotalEmitTimes.json';
 import domino from 'domino';
 import eightdiagrams from './fixtures/json/eightdiagrams.json';
 
@@ -53,6 +55,38 @@ describe('fromJSON', () => {
       proton.emitters[1].position.x,
       eightdiagrams.emitters[1].position.x
     );
+    assert.equal(
+      proton.emitters[0].rotation.x,
+      eightdiagrams.emitters[0].rotation.x
+    );
+    assert.equal(
+      proton.emitters[0].rotation.y,
+      eightdiagrams.emitters[0].rotation.y
+    );
+    assert.equal(
+      proton.emitters[0].rotation.z,
+      eightdiagrams.emitters[0].rotation.z
+    );
+
+    done();
+  });
+
+  it('should instantiate and set the total emit times and life to 1 on the emitter', done => {
+    const proton = Particles.fromJSON(containsTotalEmitTimes);
+    const emitter = proton.emitters[0];
+
+    assert.equal(emitter.totalEmitTimes, 1);
+    assert.equal(emitter.life, 1);
+
+    done();
+  });
+
+  it('should set the emitter behaviours', done => {
+    const proton = Particles.fromJSON(containsEmitterBehaviours);
+    const emitter = proton.emitters[0];
+
+    assert.notEmpty(emitter.emitterBehaviours);
+    assert.lengthOf(emitter.emitterBehaviours, 1);
 
     done();
   });