+++ CHANGELOG.md
@@ -5,6 +5,47 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## `v3.0.0` - 2019-03-01
+
+### Added
+
+- The `emitterBehaviours` array property to the `Emitter` class
+- An API to facilitate interacting with this array
+- An `updateEmitterBehaviours` method. This is called from within the main `update` method
+- An example to show emitter behaviours working
+- Tests covering new functionality
+
+### Removed
+
+- The `BehaviourEmitter` class. All functionality is now bundled into the `Emitter` class. This is a breaking change.
+
+## `v2.3.0` - 2019-02-27
+
+### Added
+
+- The ability to set mesh material blending mode correctly from a string via the `fromJSON` method for the `BodySprite` initialiser
+- Unit tests covering this
+
+## `v2.2.0` - 2019-02-17
+
+### Added
+
+- `setRotation` method to `Emitter` class
+- `rotation` prop extraction when building emitters via the `core/fromJSON` method
+- `isEnabled` property to both initializers and behaviours
+- `mutate` method to behaviours that perform internal logic
+- Test coverage for new functionalities
+
+### Changed
+
+- The behaviour `applyBehaviour` method now checks to see if the behaviour is enabled or not before applying behaviour logic to the particle/emitter via a new `mutate` method
+
+## `v2.1.0` - 2019-01-19
+
+### Added
+
+- `totalEmitTimes` and `life` JSON props for emitters
+
 ## `v2.0.0` - 2019-01-07
 
 ### Added